date 2024.7.4 v2.8
In version 2.8, we added a confirmation window for the delete item function in the UI. In the log file, we have added the username that will be changed after each record for easy viewing. In category_tree.py, a new function update_quantity has been added.
For unittest, we will continue to try to increase the coverage of the test on the code